data-processing/requirements.txt
fitparse>=1.2.0
Pillow>=10.0.0
exifread>=3.0.0
geopy>=2.3.0
python-dateutil>=2.8.0
